+++ tests/test_unit/test_openapi/test_example_snapshots.py
@@ -1,7 +1,7 @@
 import datetime as dt
 import enum
 import json
-from typing import Any
+from typing import Annotated, Any
 
 import pydantic
 import pytest
@@ -11,6 +11,7 @@ from syrupy.assertion import SnapshotAssertion
 
 from dmr import Body, Controller
 from dmr.openapi import build_schema
+from dmr.openapi.objects import Example, MediaTypeMetadata
 from dmr.plugins.pydantic import PydanticSerializer
 from dmr.routing import Router
 from dmr.settings import Settings
@@ -94,6 +95,56 @@ def test_user_schema_with_existing_examples(
                 Router(
                     'api/v1/',
                     [path('user/', _ExistingExampleController.as_view())],
+                ),
+            ).convert(),
+            indent=2,
+        )
+        == snapshot
+    )
+
+
+class _RegularBody(pydantic.BaseModel):
+    coord_x: float
+    coord_y: float
+
+
+class _ExistingBodyExamplesController(
+    Body[
+        Annotated[
+            _RegularBody,
+            MediaTypeMetadata(
+                examples={
+                    'start': Example(
+                        summary='hand written example',
+                        description='starting point',
+                        value={'coord_x': 0, 'coord_y': 0},
+                    ),
+                },
+            ),
+        ]
+    ],
+    Controller[PydanticSerializer],
+):
+    def post(self) -> int:
+        raise NotImplementedError
+
+
+def test_schema_with_body_existing_examples(
+    snapshot: SnapshotAssertion,
+    dmr_clean_settings: None,
+    settings: LazySettings,
+) -> None:
+    """Ensure that schema is correct for existing examples in body."""
+    settings.DMR_SETTINGS = {
+        Settings.openapi_examples_seed: 5,
+    }
+
+    assert (
+        json.dumps(
+            build_schema(
+                Router(
+                    'api/v1/',
+                    [path('point/', _ExistingBodyExamplesController.as_view())],
                 ),
             ).convert(),
             indent=2,