+++ CHANGELOG.md
@@ -3,6 +3,21 @@
 We follow [Semantic Versions](https://semver.org/).
 
 
+## WIP
+
+### Features
+
+- Added `MediaType` validation for the default `encoding` field
+  and OpenAPI 3.2 `itemEncoding` and `prefixEncoding` fields
+- Added `MediaTypeMetadata` metadata item to set required parameters
+  for the `MediaType` request body parameters for `Body` component
+
+### Misc
+
+- Added `$dmr-openapi-skeleton` AI agent skill, #693
+- Added `$dmr-from-django-ninja` AI agent skill, #693
+
+
 ## Version 0.3.0 (2026-03-17)
 
 ### Features